Rename repository til client, client til clientbuilder
